requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.4
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
PyJWT==1.7.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
twilio==6.49.0
urllib3==1.26.2
whitenoise==5.2.0
